requirements.txt
# TGmusicbot - Telegram 音乐管理机器人
# 功能：歌单同步 + 音乐上传 + Web 管理界面

python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
rapidfuzz==3.5.2
cryptography==41.0.7

# Web 界面
fastapi==0.109.0
uvicorn[standard]==0.25.0
jinja2==3.1.2
python-multipart==0.0.6
werkzeug==3.0.1

# 网易云音乐下载 + NCM 解密
pycryptodome==3.20.0
mutagen==1.47.0

# 文件监控（自动整理功能）
watchdog==4.0.0

# Pyrogram (大文件上传支持)
pyrogram==2.0.106
tgcrypto==1.2.5
Pillow
httpx
redis
